% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MCMC_Utilities.R
\name{is.CP}
\alias{is.CP}
\title{is.CP}
\usage{
is.CP(x)
}
\arguments{
\item{x}{object to be tested.}
}
\description{
\code{is.CP} is a general test of an object being interpretable as a
\code{\link{CP}} object.
}
\details{
The \code{\link{CP}} class is defined as the regression object that
 results from the \code{\link{CP}} regression function.
}

