% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MCMC_Utilities.R
\name{is.spCP}
\alias{is.spCP}
\title{is.spCP}
\usage{
is.spCP(x)
}
\arguments{
\item{x}{object to be tested.}
}
\description{
\code{is.spCP} is a general test of an object being interpretable as a
\code{\link{spCP}} object.
}
\details{
The \code{\link{spCP}} class is defined as the regression object that
 results from the \code{\link{spCP}} regression function.
}
