% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MCMC_Utilities_lmc.R
\name{is.spCP_lmc}
\alias{is.spCP_lmc}
\title{is.spCP_lmc}
\usage{
is.spCP_lmc(x)
}
\arguments{
\item{x}{object to be tested.}
}
\description{
\code{is.spCP_lmc} is a general test of an object being interpretable as a
\code{\link{spCP_lmc}} object.
}
\details{
The \code{\link{spCP_lmc}} class is defined as the regression object that
 results from the \code{\link{spCP_lmc}} regression function.
}

