% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MCMC_Utilities_novar.R
\name{is.spCP_novar}
\alias{is.spCP_novar}
\title{is.spCP_novar}
\usage{
is.spCP_novar(x)
}
\arguments{
\item{x}{object to be tested.}
}
\description{
\code{is.spCP_novar} is a general test of an object being interpretable as a
\code{\link{spCP_novar}} object.
}
\details{
The \code{\link{spCP_novar}} class is defined as the regression object that
 results from the \code{\link{spCP_novar}} regression function.
}

