% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PRED_predict.R
\name{predict.spCP}
\alias{predict.spCP}
\title{predict.spCP}
\usage{
\method{predict}{spCP}(object, NewTimes, ...)
}
\arguments{
\item{object}{a \code{\link{spCP}} model object for which predictions
are desired from.}

\item{NewTimes}{a numeric vector including desired time(s) points for prediction.}

\item{...}{other arguments.}
}
\value{
\code{predict.spCP} returns a list containing the following objects.

  \describe{

  \item{\code{Y}}{A \code{list} containing a matrix of predictions for each future time
  point. Each matrix has one column for each location and contains posterior
  samples obtained by Bayesian krigging.}

  }
}
\description{
Predicts future observations from the \code{\link{spCP}} model.
}
\details{
\code{predict.spCP} uses Bayesian krigging to obtain posterior samples
from future time points.
}
\author{
Samuel I. Berchuck
}
