% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reg.spCP-data.R
\docType{data}
\name{reg.spCP}
\alias{reg.spCP}
\title{Example \code{spCP} object}
\format{An object of class \code{spCP} of length 13.}
\usage{
data(reg.spCP)
}
\description{
An example \code{spCP} object obtained from the example in the documentation for the
\code{spCP} function.
}
\keyword{datasets}

